requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
aiosqlite==0.20.0
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.0

# S3 support (optional)
boto3==1.35.95

# Marker will be installed separately:
# pip install marker-pdf